+++ packages/pandas-gbq/pandas_gbq/exceptions.py
@@ -12,3 +12,12 @@ class InvalidPrivateKeyFormat(ValueError):
     """
 
     pass
+
+
+class PerformanceWarning(RuntimeWarning):
+    """
+    Raised when a performance-related feature is requested, but unsupported.
+
+    Such warnings can occur when dependencies for the requested feature
+    aren't up-to-date.
+    """